OFFSET in build macro is now optional
Parameter $OFFSET is now optional in build shell macro of guest merge.
This parameter give a placeholder for value to offset a load of code
into embedded system - usually the guest code should be flashed in different
place from compiled and linked. The offset is a calculated value to change
a code load address during creation of .SREC file.

However, in some cases the build shell macro can be multiline and $OFFSET
can be only in some of it.

Signed-off-by: Leonid Yegoshin <Leonid.Yegoshin@imgtec.com>