Microchip DM320007 board support
It is added in combination with MEBII (Microchip Extension Board II).
The reason is - console UART. On-board UART2 is connected to some funny chip
MCP2221 (USB-2-UART/I2C adapter) which _should_ be configured somehow from USB
to correctly provide UART-2-USB transmition. It also shares UART RX pin with
onboard SW3 switch and input feature can be impaired.

So, the 4-6 pins on PICTail of MEBII "motherboard" is used as UART1 for console
in provided configuration.

But nothing prevents to adjust a DM320007 board setup to use on-board UART2 as
a console.

But you can start it without console, guest blinks by color LEDs to prove
a working code. Oh, the simple guest is added as is a config file.

Note: if you change a console from UART1 to UART2 - don't forget to adjust guest
console output in appropriate way! tinyVP doesn't simulate devices, it is honest
emulator of board console/Interrupt Controller and direct access to configured
board devices.

Signed-off-by: Leonid Yegoshin <Leonid.Yegoshin@imgtec.com>