Print an ignored write to console
New option for guest device register access is added - 'I'. It is similar to
'i' - ignore write, but also logs write to that register to console.

This option is intended for accelerated guest configuration of ports/pins in
multi-guest setup.

Instead of searching in guest source code for ports/pins setup it can be looked
into console log to figure out how port/pin routing is configured by guest.
The information can be later added manually into board setup and system can be
built again with a normal 'i' and restarted.

Signed-off-by: Leonid Yegoshin <Leonid.Yegoshin@imgtec.com>