BUGFIX: mixed PTE (small and full size)
Mixed PTE case is fixed.

If two registers blocks are located in adjucent pages in pair but 1st is shorter
or incompatible then a second page should be split. Otherwise there is no
a correct access to a second page.

One example is a page with PIC32MZEF Interrupt Controller registers and
an adjuscent page with DMA controller registers.

Signed-off-by: Leonid Yegoshin <Leonid.Yegoshin@imgtec.com>