Revert "After IRQ - search for next IRQ in IC and execute it w/out ERET from interrupt"
This reverts commit 58d5a78d7dae8d58d93ddf102f859e549387f81b.

It turned out that in EIC mode MIPS CPU does provide a just happend interrupt IPL
instead of a waiting interrupt IPL in CP0 CAUSE or VIEW_RIPL.

So, there is no a clear way to know that we have another IRQ back-to-back to
a current one. The original patch doesn't work. It was found in DMA IRQ test.

Signed-off-by: Leonid Yegoshin <Leonid.Yegoshin@imgtec.com>