BUGFIX: change "current" value in compatible with new toolchain way
Variable "current" keeps a pointer to guest frame and should be changed
in a middle of context switch function. It was done via assembler inline
but it doesn't prevent a new toolchain from keeping some old values due to
optimization.

Switched to a regular C-style assigment, it seems working now.

Signed-off-by: Leonid Yegoshin <Leonid.Yegoshin@imgtec.com>
Signed-off-by: Leonid Yegoshin <Leonid.Yegoshin@mips.com>