BUGFIX: Clearing LLbit uses VM0 map but not VM7.
Small guest-mode function which clears guest LL bit on M5150 is
executed under VM7 guest. But it has CP0 CONTEXT value of VM0,
so TLB refill uses VM0 map table.

Configuration procedures are fixed to reflect it.

Besides that the new toolchain _problem_ prevents a load of correct
but complicated expression with address offset in "li" instruction.
So, it is fixed by moving that expression to memory word.

Also, a load of good values into CP0 registers is strengthened.

Signed-off-by: Leonid Yegoshin <Leonid.Yegoshin@mips.com>