SW Multithreading added
Root threads now can be defined in separate contexts. It can be any - kernel
or user mode. For user-mode thread the memory map is needed in similar way as
guest map.

Thread can be put into Shadow Register Set or share the same SRS0 with kernel.
Guest is still required to run in SRS, the difference is in treating K0/K1 GPR,
guest may use it anytime and on CPU with only two CP0 Scratch registers the IRQ
prologue can be lengthy.

Main thread now can be only kernel (not user), the point of WAIT instruction
which is a bottom of stack and is in main thread. Splitting IRQ procedure into
two, with full GPR save and partial GPR save requires main thread with reliable
GP, FP and SP registers, so it can't be user mode.

The simple kernel function "loop" is defined for easy test goal of thread.
But right now this multithreading is mostly useless until some efficient way for
inter-thread, inter-guest communication is done + some elementary service
in root for threads like "WAIT event" etc. Guest has this service via
MIPS Instruction set and CP0 Architecture and access to devices.

Some technical details:

0. Root kernel stack is now for exceptions/IRQs, root kernel threads should have
   a separate stack. Main thread 0 is exclusion.

1. The notations "_wait" and "prolog" are replaced with "in exception stack"
   (or IRQ).

2. Exception/IRQ stack context is detected now via root CP0_Status.CU0 bit of
   previous context.

3. GPRs are saved depending on SRSCtl.PSS of thread/guest - if it is zero,
   then saved on exception entry.

4. Kernel or user thread in root is detected as should be - via CP0_Status.KSU

5. Exception uses the same context (== page table tree) as thread/guest.
   It gives a possibility to examine thread/guest memory. It can be easily
   reverted to context 0 to increase guest security.

6. Each thread has it's own ASID.

7. Clear guest LLbit funcition uses now a thread 0 page table tree.
   This function may be removed in future for inter-guest, inter-thread LL/SC
   synchronization.

Signed-off-by: Leonid Yegoshin <Leonid.Yegoshin@mips.com>